Split Video Populate Data Steps
Only populating the validHues vector when the query hues match the src hues so we don't have to load in all the frames for every single src video. This reduces the search time by around 40% (5s->3s for video_11), and the time save becomes more significant the more videos the search has to look through.